+++ saleor/graphql/discount/mutations/sale/sale_remove_catalogues.py
@@ -2,7 +2,7 @@ from typing import cast
 
 from .....core.tracing import traced_atomic_transaction
 from .....discount import models
-from .....discount.models import Promotion
+from .....discount.sale_converter import get_or_create_promotion
 from .....discount.utils import fetch_catalogue_info
 from .....graphql.channel import ChannelContext
 from .....permission.enums import DiscountPermissions
@@ -40,7 +40,7 @@ class SaleRemoveCatalogues(SaleBaseCatalogueMutation):
             cls.get_node_or_error(info, id, only_type=Sale, field="sale_id"),
         )
         previous_catalogue = fetch_catalogue_info(sale)
-        promotion = Promotion.objects.get(old_sale_id=sale.id)
+        promotion = get_or_create_promotion(sale)
         rules = promotion.rules.all()
         manager = get_plugin_manager_promise(info.context).get()
         with traced_atomic_transaction():
