+++ saleor/graphql/discount/tests/mutations/test_sale_catalogues_remove.py
@@ -3,13 +3,10 @@ from unittest.mock import patch
 import graphene
 
 from .....discount.models import Promotion
-from .....discount.sale_converter import (
-    convert_migrated_sale_predicate_to_catalogue_info,
-    create_promotion_for_new_sale,
-)
 from .....discount.utils import fetch_catalogue_info
 from ....tests.utils import get_graphql_content
 from ...mutations.utils import convert_catalogue_info_to_global_ids
+from .utils import convert_migrated_sale_predicate_to_catalogue_info
 
 SALE_CATALOGUES_REMOVE_MUTATION = """
     mutation saleCataloguesRemove($id: ID!, $input: CatalogueInput!) {
@@ -60,7 +57,6 @@ def test_sale_remove_catalogues(
         graphene.Node.to_global_id("ProductVariant", variant.id)
         for variant in product_variant_list
     ]
-    create_promotion_for_new_sale(sale, previous_catalogue)
     variables = {
         "id": graphene.Node.to_global_id("Sale", sale.id),
         "input": {
